+++ R/sjPlotGLME.R
@@ -270,7 +270,7 @@ sjp.glmer <- function(fit,
                       prnt.plot = TRUE,
                       ...) {
 
-  if (stats::runif(1) < .3)
+  if (stats::runif(1) < .35)
     message("`sjp.glmer()` will become deprecated in the future. Please use `plot_model()` instead.")
 
   ## TODO activate in future update
@@ -577,7 +577,7 @@ sjp.lmer <- function(fit,
                      prnt.plot = TRUE,
                      ...) {
 
-  if (stats::runif(1) < .3)
+  if (stats::runif(1) < .35)
     message("`sjp.lmer()` will become deprecated in the future. Please use `plot_model()` instead.")
 
   ## TODO activate in future update