+++ R/sjPlotLinreg.R
@@ -296,7 +296,7 @@ sjp.lm <- function(fit,
                    prnt.plot = TRUE,
                    ...) {
 
-  if (stats::runif(1) < .3)
+  if (stats::runif(1) < .35)
     message("`sjp.lm()` will become deprecated in the future. Please use `plot_model()` instead.")
 
   ## TODO activate in future update